+++ libs/ionic-react/src/utils/versions.ts
@@ -1,9 +1,9 @@
 export const nxtendVersion = '*';
 
-export const ionicReactVersion = '^5.2.2';
-export const ionicReactRouterVersion = '^5.2.2';
+export const ionicReactVersion = '5.2.2';
+export const ionicReactRouterVersion = '5.2.2';
 export const ioniconsVersion = '5.0.1';
 
-export const testingLibraryJestDomVersion = '5.5.0';
-export const testingLibraryUserEventVersion = '10.0.1';
+export const testingLibraryJestDomVersion = '5.11.0';
+export const testingLibraryUserEventVersion = '12.0.11';
 export const testingLibraryCypressVersion = '6.0.0';